requirements.txt
Flask==2.3.0
tensorflow==2.16.1
joblib==1.2.0
numpy==1.26.4
scikit-learn==1.3.2
pandas==2.2.2
minimalmodbus==2.0.1
pyserial==3.5